Exercise 3.20
Add validation to your phonebook application, that will make sure that phone numbers are of the correct form. A phone number must

has length of 8 or more
if formed of two parts that are separated by -, the first part has two or three numbers and the second part also consists of numbers

eg. 09-1234556 and 040-22334455 are valid phone numbers
eg. 1234556, 1-22334455 and 10-22-334455 are invalid
Use a Custom validator to implement the second part of the validation.

If an HTTP POST request tries to add a name that is already in the phonebook, the server must respond with an appropriate status code and error message.